requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Background Tasks
rq==1.15.1
redis==5.0.1

# Security & Validation
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6

# Rate Limiting
slowapi==0.1.9

# Monitoring & Logging
prometheus-client==0.19.0
loguru==0.7.2

# System Monitoring
psutil==5.9.6

# Audio Processing
yt-dlp==2023.11.16
demucs==4.0.1
torch==2.1.1
torchaudio==2.1.1
ffmpeg-python==0.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1